requirements.txt
timm
opencv-python
tensorflow
pillow
torch 
torchvision
numpy
